fix AST::substitute
