+++ include/cfir/Printer.h
@@ -11,4 +11,3 @@
 void print_type_checker_condition(std::ostream &stream, const std::shared_ptr<CFIR::Identifier> &current_id, const std::string &type_name, const std::shared_ptr<CFIR::Identifier> &output_id);
 
 IdPtr make_new_unique_name();
-AST::ExprPtr substitute(const AST::ExprPtr &expr, const VarScope &scope);
